Inline interface resources and simplify inheritance
