+++ src/entity/resources/__init__.py
@@ -3,7 +3,9 @@ from .llm import LLM
 from .memory import Memory
 from .storage import Storage
 from .metrics import MetricsCollectorResource
-from .interfaces import DatabaseResource, LLMResource, VectorStoreResource
+from .database import DatabaseResource
+from .vector_store import VectorStoreResource
+from .llm import LLMResource
 
 __all__ = [
     "AgentResource",
