+++ src/entity/resources/duckdb_vector_store.py
