--- src/entity/resources/interfaces/__init__.py
@@ -1,19 +1,0 @@
-from .database import DatabaseResource
-from .duckdb_resource import DuckDBResource
-from .postgres_resource import PostgresResource
-from .duckdb_vector_store import DuckDBVectorStore
-from .echo_llm import EchoLLMResource
-from .llm import LLMResource
-from .storage import StorageResource
-from .vector_store import VectorStoreResource
-
-__all__ = [
-    "DatabaseResource",
-    "VectorStoreResource",
-    "DuckDBVectorStore",
-    "DuckDBResource",
-    "PostgresResource",
-    "EchoLLMResource",
-    "LLMResource",
-    "StorageResource",
-]
