+++ tests/test_reload_unknown_plugin.py
@@ -7,8 +7,8 @@ from entity.core.plugins import Plugin, ValidationResult
 from entity.core.stages import PipelineStage
 from entity.cli import EntityCLI
 from entity.infrastructure import DuckDBInfrastructure
-from entity.resources.interfaces.duckdb_resource import DuckDBResource
-from entity.resources.interfaces.duckdb_vector_store import DuckDBVectorStore
+from entity.resources.database import DuckDBResource
+from entity.resources.duckdb_vector_store import DuckDBVectorStore
 
 from .test_reload_runtime_validation import run_reload
