+++ tests/test_resources/test_memory_init.py
@@ -1,7 +1,7 @@
 import pytest
 from entity.resources import Memory
-from entity.resources.interfaces.vector_store import VectorStoreResource
-from entity.resources.interfaces.database import DatabaseResource
+from entity.resources.vector_store import VectorStoreResource
+from entity.resources.database import DatabaseResource
 from entity.pipeline.errors import ResourceInitializationError
 from contextlib import asynccontextmanager
